app/requirement.txt
passlib
motor
python-jose
fastapi["standard]
bcrypt